Shrink Docker image
Migrate this image from the python:3.5 base image to the much smaller
python:3.6-alpine base image.

To bring this container up to date with the currently established best
practices, the web serving infrastructure is swapped from NGINX with
uwsgi, to gunicorn, with the whitenoise library for serving static
files.

This reduces the container image size from 844 MByte to 155 MByte (an
82% reduction).