Add extended site information
Site objects are extended with the following fields:

- lat -- latitude of the site as a whole
- long -- longitude of the site as a whole
- tz -- the name of the timezone where the site resides
- restart -- the UTC time at which the site restarts

These additional field are useful for a significant number of tools in
the LCO universe. With this information residing in a central location,
these tools can start consuming consistent information without having to
duplicate it in each and every project.

- LCO Timezone Helper (http://lcotime.lco.gtn/)
- Puppet
- Allsky Camera Day/Night scheduler
- Calibrations project
- NetOps Calendar
- LCO Telescope Operations Slack Bot
- Monitoring Systems (Nagios, etc.)

The timezone name is needed, because the existing timezone field is
incorrect in many timezones where LCO operates telescopes, for at least
some part of each year. Any tools which care about accurate timezone
information will be using the timezone name to calculate the current
offset at the date they require.

The UTC offset of most timezones changes several times per year. In
addition, the UTC offset of several timezones around the world is
measured in minutes, not hours. As an example, see Australia/Adelaide
(where Andrew Pickles resides) which has a +09:30 offset at the time of
this commit.